+++ pkgs/unified_analytics/lib/src/config_handler.dart
@@ -19,9 +19,9 @@ const String telemetryFlagPattern = r'^reporting=([0|1]) *$';
 /// from the configuration file
 ///
 /// Example:
-/// flutter_tools=2022-10-26,1
+/// flutter-tool=2022-10-26,1
 const String toolPattern =
-    r'^([A-Za-z0-9]+[A-Za-z0-9_]*)=([0-9]{4}-[0-9]{2}-[0-9]{2}),([0-9]+)$';
+    r'^([A-Za-z0-9]+[A-Za-z0-9_-]*)=([0-9]{4}-[0-9]{2}-[0-9]{2}),([0-9]+)$';
 
 class ConfigHandler {
   /// Regex pattern implementation for matching a line in the config file