+++ pkgs/unified_analytics/lib/src/enums.dart
@@ -19,44 +19,37 @@ enum DashEvent {
   hotReloadTime(
     label: 'hot_reload_time',
     description: 'Hot reload duration',
-    toolOwner: DashTool.flutterTools,
+    toolOwner: DashTool.flutterTool,
   ),
 
   // Events for language_server
   clientNotification(
     label: 'client_notification',
     description: 'Notifications sent from the client',
-    toolOwner: DashTool.languageServer,
   ),
   clientRequest(
     label: 'client_request',
     description: 'Requests sent from the client',
-    toolOwner: DashTool.languageServer,
   ),
   contextStructure(
     label: 'context_structure',
     description: 'Structure of the analysis contexts being analyzed',
-    toolOwner: DashTool.languageServer,
   ),
   lintUsageCounts(
     label: 'lint_usage_counts',
     description: 'Number of times each lint is enabled',
-    toolOwner: DashTool.languageServer,
   ),
   pluginRequest(
     label: 'plugin_request',
     description: 'Request responses from plugins',
-    toolOwner: DashTool.languageServer,
   ),
   serverSession(
     label: 'server_session',
     description: 'Dart Analyzer Server session data',
-    toolOwner: DashTool.languageServer,
   ),
   severityAdjustments(
     label: 'severity_adjustments',
     description: 'Number of times diagnostic severity is changed',
-    toolOwner: DashTool.languageServer,
   ),
   ;
 
@@ -70,25 +63,53 @@ enum DashEvent {
   });
 }
 
-/// Officially-supported clients of this package.
-///
-/// All [label] values should use an underscore as a delimiter.
+/// Officially-supported clients of this package as logical
+/// tools, grouped by user point of view.  Derived directly
+/// from the PDD.
 enum DashTool {
+  androidStudioPlugins(
+    label: 'android-studio-plugins',
+    description: 'Android Studio IDE plugins for Dart and Flutter',
+  ),
+  dartTool(
+    label: 'dart-tool',
+    description: 'Dart CLI developer tool',
+  ),
+  @Deprecated('From a prerelease version of analytics; do not use.')
   dartTools(
     label: 'dart_tools',
     description: 'A CLI for Dart development',
   ),
+  devtools(
+    label: 'devtools',
+    description: 'DevTools debugging and performance tools',
+  ),
+  flutterTool(
+    label: 'flutter-tool',
+    description: 'Flutter CLI developer tool',
+  ),
+  @Deprecated('From a prerelease version of analytics; do not use.')
   flutterTools(
     label: 'flutter_tools',
     description: 'Runs flutter applications from CLI',
   ),
-  languageServer(
-    label: 'language_server',
-    description: 'The Dart language server for IDE and CLI support.',
+  intellijPlugins(
+    label: 'intellij-plugins',
+    description: 'IntelliJ IDE plugins for Dart and Flutter',
+  ),
+  vscodePlugins(
+    label: 'vscode-plugins',
+    description: 'VS Code IDE plugins for Dart and Flutter',
   );
 
+  /// String used as the control flag and the value of the tool key in
+  /// analytics.
   final String label;
+
+  /// The "notice string", a human-readable description of the logical tool
+  /// grouping.
   final String description;
+
   const DashTool({
     required this.label,
     required this.description,