+++ pkgs/unified_analytics/test/unified_analytics_test.dart
@@ -31,8 +31,8 @@ void main() {
   late UserProperty userProperty;
 
   const String homeDirName = 'home';
-  const DashTool initialTool = DashTool.flutterTools;
-  const DashTool secondTool = DashTool.dartTools;
+  const DashTool initialTool = DashTool.flutterTool;
+  const DashTool secondTool = DashTool.dartTool;
   const String measurementId = 'measurementId';
   const String apiSecret = 'apiSecret';
   const int toolsMessageVersion = 1;
@@ -1053,7 +1053,7 @@ ${initialTool.label}=${ConfigHandler.dateStamp},$toolsMessageVersion
       // Checks item 3
       // All tools must be under 36 characters (and enforce each tool
       // begins with a letter)
-      final RegExp toolLabelPattern = RegExp(r'^[a-zA-Z][a-zA-Z\_]{0,35}$');
+      final RegExp toolLabelPattern = RegExp(r'^[a-zA-Z][a-zA-Z\_-]{0,35}$');
       bool toolLengthValid = true;
       final List<DashTool> invalidTools = <DashTool>[];
       for (DashTool tool in DashTool.values) {
@@ -1099,9 +1099,18 @@ ${initialTool.label}=${ConfigHandler.dateStamp},$toolsMessageVersion
     // Retrieve the consent message for flutter tools
     final String consentMessage = analytics.getConsentMessage;
 
-    expect(
-        consentMessage,
-        kToolsMessage.replaceAll(
-            '[tool name]', DashTool.flutterTools.label.replaceAll('_', ' ')));
+    expect(consentMessage, equalsIgnoringWhitespace(r'''
+The Flutter CLI developer tool uses Google Analytics to report usage and
+diagnostic data along with package dependencies, and crash reporting to
+send basic crash reports. This data is used to help improve the Dart platform,
+Flutter framework, and related tools. Telemetry is not sent on the very first
+run. To disable reporting of telemetry, run this terminal command:
+
+[dart|flutter] --disable-telemetry.
+
+If you opt out of telemetry, an opt-out event will be sent, and then no
+further information will be sent. This data is collected in accordance with
+the Google Privacy Policy (https://policies.google.com/privacy).
+'''));
   });
 }