+++ pkgs/unified_analytics/test/workflow_test.dart
@@ -22,8 +22,8 @@ void main() {
   late File logFile;
 
   const String homeDirName = 'home';
-  const DashTool initialTool = DashTool.flutterTools;
-  const DashTool secondTool = DashTool.dartTools;
+  const DashTool initialTool = DashTool.flutterTool;
+  const DashTool secondTool = DashTool.dartTool;
   const String measurementId = 'measurementId';
   const String apiSecret = 'apiSecret';
   const int toolsMessageVersion = 1;