+++ pkgs/unified_analytics/CHANGELOG.md
@@ -1,7 +1,8 @@
-## 0.1.3-dev
+## 1.0.0-dev
 
 - Error handling functionality added to prevent malformed session json data from causing a crash
 - Creating a new analytics constructor to point to a test instance of Google Analytics for developers
+- Align supported tool list with PDD
 - Exposing a new instance method that will need to be invoked when a client has successfully shown the consent message to the user `clientShowedMessage()`
 - Adding and incrementing a tool's version will automatically use the current consent message version instead of incrementing by 1